fix: remove fallback to current block height
When starting a new node, if we cannot get the information about fromHeight
block possibly due to flaky RPC, we fallbacks to get the latest block. This can
make the operator assume that all blocks from fromHeight to latest block have
been processed. This commit removes the fallback and always retries to get the
fromHeight block.